Feat: Custom hooks added
